+++ src/Symfony/Bundle/FrameworkBundle/Console/Descriptor/MarkdownDescriptor.php
@@ -183,6 +183,7 @@ class MarkdownDescriptor extends Descriptor
             ."\n".'- Shared: '.($definition->isShared() ? 'yes' : 'no')
             ."\n".'- Abstract: '.($definition->isAbstract() ? 'yes' : 'no')
             ."\n".'- Autowired: '.($definition->isAutowired() ? 'yes' : 'no')
+            ."\n".'- Autoconfigured: '.($definition->isAutoconfigured() ? 'yes' : 'no')
         ;
 
         foreach ($definition->getAutowiringTypes(false) as $autowiringType) {
@@ -343,7 +344,8 @@ class MarkdownDescriptor extends Descriptor
         }
 
         if ($callable instanceof \Closure) {
-            $string .= "\n- Type: `closure`";
+            $formatted = $this->formatClosure($callable);
+            $string .= "\n- Type: `$formatted`";
 
             return $this->write($string."\n");
         }