+++ src/Symfony/Component/DependencyInjection/Loader/XmlFileLoader.php
@@ -180,6 +180,9 @@ class XmlFileLoader extends FileLoader
         if ($defaultsNode->hasAttribute('inherit-tags')) {
             $defaults['inherit-tags'] = XmlUtils::phpize($defaultsNode->getAttribute('inherit-tags'));
         }
+        if ($defaultsNode->hasAttribute('autoconfigure')) {
+            $defaults['autoconfigure'] = XmlUtils::phpize($defaultsNode->getAttribute('autoconfigure'));
+        }
 
         return $defaults;
     }
@@ -222,12 +225,22 @@ class XmlFileLoader extends FileLoader
             $defaults = array();
         } else {
             $definition = new Definition();
+
+            if (isset($defaults['public'])) {
+                $definition->setPublic($defaults['public']);
+            }
+            if (isset($defaults['autowire'])) {
+                $definition->setAutowired($defaults['autowire']);
+            }
+            if (isset($defaults['autoconfigure'])) {
+                $definition->setAutoconfigured($defaults['autoconfigure']);
+            }
+
+            $definition->setChanges(array());
         }
 
         if ($publicAttr = $service->getAttribute('public')) {
             $definition->setPublic(XmlUtils::phpize($publicAttr));
-        } elseif (isset($defaults['public'])) {
-            $definition->setPublic($defaults['public']);
         }
 
         foreach (array('class', 'shared', 'synthetic', 'lazy', 'abstract') as $key) {
@@ -239,8 +252,10 @@ class XmlFileLoader extends FileLoader
 
         if ($value = $service->getAttribute('autowire')) {
             $definition->setAutowired(XmlUtils::phpize($value));
-        } elseif (isset($defaults['autowire'])) {
-            $definition->setAutowired($defaults['autowire']);
+        }
+
+        if ($value = $service->getAttribute('autoconfigure')) {
+            $definition->setAutoconfigured(XmlUtils::phpize($value));
         }
 
         if ($files = $this->getChildren($service, 'file')) {