+++ src/Symfony/Component/DependencyInjection/Loader/YamlFileLoader.php
@@ -57,6 +57,7 @@ class YamlFileLoader extends FileLoader
         'decoration_priority' => 'decoration_priority',
         'autowire' => 'autowire',
         'autowiring_types' => 'autowiring_types',
+        'autoconfigure' => 'autoconfigure',
     );
 
     private static $prototypeKeywords = array(
@@ -75,21 +76,19 @@ class YamlFileLoader extends FileLoader
         'tags' => 'tags',
         'inherit_tags' => 'inherit_tags',
         'autowire' => 'autowire',
+        'autoconfigure' => 'autoconfigure',
     );
 
     private static $instanceofKeywords = array(
         'shared' => 'shared',
         'lazy' => 'lazy',
         'public' => 'public',
-        'abstract' => 'abstract',
-        'deprecated' => 'deprecated',
-        'factory' => 'factory',
-        'arguments' => 'arguments',
         'properties' => 'properties',
         'configurator' => 'configurator',
         'calls' => 'calls',
         'tags' => 'tags',
         'autowire' => 'autowire',
+        'autoconfigure' => 'autoconfigure',
     );
 
     private static $defaultsKeywords = array(
@@ -97,6 +96,7 @@ class YamlFileLoader extends FileLoader
         'tags' => 'tags',
         'inherit_tags' => 'inherit_tags',
         'autowire' => 'autowire',
+        'autoconfigure' => 'autoconfigure',
     );
 
     private $yamlParser;
@@ -366,6 +366,18 @@ class YamlFileLoader extends FileLoader
             $defaults = array();
         } else {
             $definition = new Definition();
+
+            if (isset($defaults['public'])) {
+                $definition->setPublic($defaults['public']);
+            }
+            if (isset($defaults['autowire'])) {
+                $definition->setAutowired($defaults['autowire']);
+            }
+            if (isset($defaults['autoconfigure'])) {
+                $definition->setAutoconfigured($defaults['autoconfigure']);
+            }
+
+            $definition->setChanges(array());
         }
 
         if (isset($service['class'])) {
@@ -384,9 +396,8 @@ class YamlFileLoader extends FileLoader
             $definition->setLazy($service['lazy']);
         }
 
-        $public = isset($service['public']) ? $service['public'] : (isset($defaults['public']) ? $defaults['public'] : null);
-        if (null !== $public) {
-            $definition->setPublic($public);
+        if (isset($service['public'])) {
+            $definition->setPublic($service['public']);
         }
 
         if (isset($service['abstract'])) {
@@ -484,9 +495,8 @@ class YamlFileLoader extends FileLoader
             $definition->setDecoratedService($service['decorates'], $renameId, $priority);
         }
 
-        $autowire = isset($service['autowire']) ? $service['autowire'] : (isset($defaults['autowire']) ? $defaults['autowire'] : null);
-        if (null !== $autowire) {
-            $definition->setAutowired($autowire);
+        if (isset($service['autowire'])) {
+            $definition->setAutowired($service['autowire']);
         }
 
         if (isset($service['autowiring_types'])) {
@@ -505,6 +515,10 @@ class YamlFileLoader extends FileLoader
                     $definition->addAutowiringType($autowiringType);
                 }
             }
+        }
+
+        if (isset($service['autoconfigure'])) {
+            $definition->setAutoconfigured($service['autoconfigure']);
         }
 
         if (array_key_exists('resource', $service)) {