+++ src/Symfony/Component/Form/Tests/DependencyInjection/FormPassTest.php
@@ -164,21 +164,52 @@ class FormPassTest extends TestCase
     /**
      * @dataProvider privateTaggedServicesProvider
      */
-    public function testPrivateTaggedServices($id, $tagName, array $tagAttributes = array())
+    public function testPrivateTaggedServices($id, $tagName, callable $assertion, array $tagAttributes = array())
     {
-        $container = $this->createContainerBuilder();
+        $formPass = new FormPass();
+        $container = new ContainerBuilder();
 
         $container->setDefinition('form.extension', $this->createExtensionDefinition());
         $container->register($id, 'stdClass')->setPublic(false)->addTag($tagName, $tagAttributes);
-        $container->compile();
+        $formPass->process($container);
+
+        $assertion($container);
     }
 
     public function privateTaggedServicesProvider()
     {
         return array(
-            array('my.type', 'form.type'),
-            array('my.type_extension', 'form.type_extension', array('extended_type' => 'Symfony\Component\Form\Extension\Core\Type\FormType')),
-            array('my.guesser', 'form.type_guesser'),
+            array(
+                'my.type',
+                'form.type',
+                function (ContainerBuilder $container) {
+                    $formTypes = $container->getDefinition('form.extension')->getArgument(0);
+
+                    $this->assertInstanceOf(Reference::class, $formTypes);
+
+                    $locator = $container->getDefinition((string) $formTypes);
+                    $expectedLocatorMap = array(
+                        'stdClass' => new ServiceClosureArgument(new Reference('my.type')),
+                    );
+
+                    $this->assertInstanceOf(Definition::class, $locator);
+                    $this->assertEquals($expectedLocatorMap, $locator->getArgument(0));
+                },
+            ),
+            array(
+                'my.type_extension',
+                'form.type_extension',
+                function (ContainerBuilder $container) {
+                    $this->assertEquals(
+                        array('Symfony\Component\Form\Extension\Core\Type\FormType' => new IteratorArgument(array(new Reference('my.type_extension')))),
+                        $container->getDefinition('form.extension')->getArgument(1)
+                    );
+                },
+                array('extended_type' => 'Symfony\Component\Form\Extension\Core\Type\FormType'),
+            ),
+            array('my.guesser', 'form.type_guesser', function (ContainerBuilder $container) {
+                $this->assertEquals(new IteratorArgument(array(new Reference('my.guesser'))), $container->getDefinition('form.extension')->getArgument(2));
+            }),
         );
     }